remove obsolete standalone test
